cloud-functions/model-retrainer/requirements.txt
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.38.0

